+++ lib/traceroute/sockets/udp.ex
@@ -57,7 +57,7 @@ defmodule Traceroute.Sockets.UDP do
   def send(packet, ip, ttl, timeout, ip_protocol, opts \\ []) do
     args = [packet: packet, ip: ip, ttl: ttl, timeout: timeout, ip_protocol: ip_protocol] ++ opts
     {:ok, pid} = start_link(args)
-    GenServer.call(pid, :send_probe, :infinity)
+    Utils.safe_genserver_call(pid, :send_probe, to_timeout(second: timeout + 1))
   end
 
   def start_link(args) do
@@ -88,9 +88,15 @@ defmodule Traceroute.Sockets.UDP do
 
   @impl GenServer
   def handle_call(:send_probe, from, state) do
-    %{domain: domain, protocol: protocol, ttl_opt: ttl_opt} = Utils.get_protocol_options(state.ip_protocol, :udp)
+    %{
+      domain: domain,
+      protocol: protocol,
+      ttl_opt: ttl_opt,
+      socket_type: socket_type
+    } =
+      Utils.get_protocol_options(state.ip_protocol, :udp)
 
-    with {:ok, udp_socket} <- :socket.open(domain, :dgram, protocol),
+    with {:ok, udp_socket} <- :socket.open(domain, socket_type, protocol),
          :ok <- :socket.setopt(udp_socket, ttl_opt, state.ttl),
          # Bind to port 0 to let the OS assign an ephemeral port
          :ok <- :socket.bind(udp_socket, %{family: domain, addr: Utils.any_addr(domain), port: 0}),